+++ src/main/kotlin/cz/cvut/fel/omo/foodchain/Foodchain/parties/Farmer.kt
@@ -2,7 +2,7 @@ package cz.cvut.fel.omo.foodchain.Foodchain.parties
 
 import cz.cvut.fel.omo.foodchain.Foodchain.products.Crop
 
-class Farmer(subjectName : String, identier : Int, location : String, amountOfMoney : Int)
+class Farmer(subjectName : String, identier : Int, location : String, amountOfMoney : Double)
     : BaseParty(subjectName, identier, location, amountOfMoney) {
 
     var resources : List<Crop> = setInitialResources()
