+++ src/main/kotlin/cz/cvut/fel/omo/foodchain/Foodchain/parties/Regulator.kt
@@ -1,5 +1,5 @@
 package cz.cvut.fel.omo.foodchain.Foodchain.parties
 
-class Regulator(subjectName : String, identier : Int, location : String, amountOfMoney : Int)
+class Regulator(subjectName : String, identier : Int, location : String, amountOfMoney : Double)
     : BaseParty(subjectName, identier, location, amountOfMoney) {
 }
