+++ src/main/kotlin/cz/cvut/fel/omo/foodchain/Foodchain/parties/Retailer.kt
@@ -1,6 +1,6 @@
 package cz.cvut.fel.omo.foodchain.Foodchain.parties
 
-class Retailer(subjectName : String, identier : Int, location : String, amountOfMoney : Int)
+class Retailer(subjectName : String, identier : Int, location : String, amountOfMoney : Double)
     : BaseParty(subjectName, identier, location, amountOfMoney) {
 }
